Improve device-side of pick display

+++ src/com/gadgetworks/codeshelf/device/LedPos.java
@@ -1,7 +1,7 @@
 /*******************************************************************************
  *  CodeShelf
  *  Copyright (c) 2005-2013, Jeffrey B. Williams, All rights reserved
- *  $Id: LedPos.java,v 1.2 2013/04/02 04:29:02 jeffw Exp $
+ *  $Id: LedPos.java,v 1.3 2013/04/14 23:33:18 jeffw Exp $
  *******************************************************************************/
 package com.gadgetworks.codeshelf.device;
 
@@ -24,11 +24,11 @@ public class LedPos {
 
 	private static final Logger	LOGGER	= LoggerFactory.getLogger(LedPos.class);
 
-	private Integer				mPosNum = 0;
+	private Short				mPosNum = 0;
 	private Integer				mCycleNum = 0;
 	private List<LedValue>		mValues;
 
-	public LedPos(final Integer inPosNum) {
+	public LedPos(final Short inPosNum) {
 		mPosNum = inPosNum;
 		mValues = new ArrayList<LedValue>();
 	}
